Move tektronix tests to different file
